+++ docs/dev_guide/changelog.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v1.4.1
+### Fixed
+- Fixed errors using pandas>=3
+
 ## v1.4.0
 ### Added
 - `huracanpy.info.beaufort_category`